# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'http://localhost:3000'
  - url: 'https://dev.iop.agetic.gob.bo/backend-base'
  - url: 'https://base.agcs.agetic.gob.bo/ws'
paths:
...
  /api/usuarios/{id}/activacion:
    patch:
      summary: Activa un usuario
      tags: [ 'Usuario' ]
      operationId: get-api-v1-usuarios-activacion
      description: Api para activar un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: ID de usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: '¡¡Registro actualizado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'ACTIVO'
  /api/usuarios/{id}/inactivacion:
    patch:
      summary: Inactiva un usuario
      tags: [ 'Usuario' ]
      operationId: get-api-v1-usuarios-inactivacion
      description: Api para inactivar un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: ID de usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: '¡¡Registro actualizado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'INACTIVO'
  /api/usuarios/cuenta/contrasena:
    patch:
      summary: Actualiza la contrasena de un usuario autenticado
      tags: [ 'Usuario' ]
      operationId: get-api-v1-usuarios-activacion-id-contrasena
      description: Api para actualizar la contrasena de un usuario
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contrasenaActual:
                  type: string
                  example: 'contrasena'
                contrasenaNueva:
                  type: string
                  example: 'contrasena-nueva'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: '¡Registro actualizado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'ACTIVO'
  /api/usuarios/cuenta/desbloqueo:
    get:
      summary: Desbloquea una cuenta bloqueada por muchos intentos fallidos
      tags: [ 'Usuario' ]
      operationId: get-api-v1-usuarios-cuenta-desbloqueo
      description: Api para desbloquear una cuenta
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: id de desbloqueo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Cuenta desbloqueada exitosamente.'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
  /api/usuarios/{id}/restauracion:
    patch:
      summary: Restaura la contrasena de un usuario
      tags: [ 'Usuario' ]
      operationId: get-api-v1-usuarios-contrasena-id
      description: Api para restaurar la contrasena de un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: ID de usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Restauración de contraseña exitosa'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'PENDIENTE'
...
  /api/usuarios:
    get:
      summary: Api para obtener el listado de usuarios
      tags: [ 'Usuario' ]
      operationId: get-api-usuarios
      description: Api para obtener el listado de usuarios
      parameters:
        - in: query
          name: pagina
          schema:
            type: string
            example: 1
          description: numero de la pagina a visualizar
        - in: query
          name: limite
          schema:
            type: string
            example: 10
          description: numero de filas a visualizar
        - in: query
          name: orden
          schema:
            type: string
            example: ASC
          description: Tipo de orden a visualizar
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro(s) obtenido(s) con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      filas:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                            usuario:
                              type: string
                              example: 'ADMINISTRADOR'
                            estado:
                              type: string
                              example: 'ACTIVO'
                            ciudadaniaDigital:
                              type: boolean
                              example: false
                            persona:
                              type: object
                              properties:
                                nroDocumento:
                                  type: string
                                  example: '1234567'
                                nombres:
                                  type: string
                                  example: 'JUAN'
                                primerApellido:
                                  type: string
                                  example: 'PEREZ'
                                segundoApellido:
                                  type: string
                                  example: 'PEREZ'
                            usuarioRol:
                              type: array
                              items:
                                type: object
                                properties:
                                  estado:
                                    type: string
                                    example: 'ACTIVO'
                                  id:
                                    type: string
                                    example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                                  rol:
                                    type: object
                                    properties:
                                      rol:
                                        type: string
                                        example: 'ADMINISTRADOR'
                      total:
                        type: number
                        example: 1
    post:
      summary: Api para crear un nuevo usuario
      tags: [ 'Usuario' ]
      operationId: post-api-usuarios
      description: Api para crear un nuevo usuario
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                correoElectronico:
                  type: string
                  example: '123456@yopmail.com'
                persona:
                  type: object
                  properties:
                    nombres:
                      type: string
                      example: 'JUAN'
                    primerApellido:
                      type: string
                      example: 'PEREZ'
                    segundoApellido:
                      type: string
                      example: 'PEREZ'
                    nroDocumento:
                      type: string
                      example: '123456'
                    fechaNacimiento:
                      type: string
                      example: '1999-11-11'
                roles:
                  type: array
                  items:
                    type: string
                  example: [ 'd5de12df-3cc3-5a58-a742-be24030482d8' ]
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: '81fd2f36-2857-41e5-aa7f-a8101297337d'
                      estado:
                        type: string
                        example: 'CREADO'
  /api/usuarios/cuenta/ciudadania:
    post:
      summary: Api para crear un nuevo usuario relacionado con Ciudadanía Digital
      tags: [ 'Usuario' ]
      operationId: post-api-usuarios-ciudadania
      description: Api para crear un nuevo usuario relacionado con Ciudadanía Digital
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                  example: '123456'
                roles:
                  type: array
                  items:
                    type: string
                  example: [ 'd5de12df-3cc3-5a58-a742-be24030482d8' ]
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: '81fd2f36-2857-41e5-aa7f-a8101297337d'
                      estado:
                        type: string
                        example: 'CREADO'
  /api/usuarios/{id}:
    patch:
      summary: Actualiza datos de un usuario
      tags: [ 'Usuario' ]
      operationId: patch-api-v1-usuarios-id
      description: Api para actualizar datos de un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: ID de usuario
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                correoElectronico:
                  type: string
                  example: 'correo@yopmail.com'
                roles:
                  type: array
                  items:
                    type: string
                  example: [ 'd5de12df-3cc3-5a58-a742-be24030482d8' ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: '¡Registro actualizado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
...